Have remove migrate:fresh from production

+++ app/Console/Commands/StartUp.php
@@ -41,7 +41,10 @@ class StartUp extends Command
      */
     public function handle()
     {
-        Artisan::call('migrate:fresh');
+        if (config('app.env') === "local"){
+            Artisan::call('migrate:fresh');
+        }
+
 
 
         $this->comment("Creating Super Admin");
